service/Services/entities/src/main/java/org/infy/idp/entities/jobs/testinfo/TestEnv.java
/***********************************************************************************************
*
* Copyright 2018 Infosys Ltd.
* Use of this source code is governed by MIT license that can be found in the LICENSE file or at
* https://opensource.org/licenses/MIT.
*
***********************************************************************************************/

package org.infy.idp.entities.jobs.testinfo;

import java.util.List;

import com.google.gson.annotations.Expose;
import com.google.gson.annotations.SerializedName;

/**
 * Entity to store test step list
 * 
 * @author Infosys
 *
 */

public class TestEnv {

	@SerializedName("envName")
	@Expose
	private String envName;
	@SerializedName("envFlag")
	@Expose
	private String envFlag;

	@SerializedName("testSteps")
	@Expose
	private List<TestStep> testSteps = null;

	public String getEnvFlag() {
		return envFlag;
	}

	public void setEnvFlag(String envFlag) {
		this.envFlag = envFlag;
	}

	public String getEnvName() {
		return envName;
	}

	public void setEnvName(String envName) {
		this.envName = envName;
	}

	public List<TestStep> getTestSteps() {
		return testSteps;
	}

	public void setTestSteps(List<TestStep> testSteps) {
		this.testSteps = testSteps;
	}

}
